Add support for compiling to WASM
Instead of calling functions from math.h, use the compiler's builtin
functions for those that correspond to WASM instructions. Also, implement
fmin and fmax with isless and isgreater respectively, since there is no
corresponding WASM instruction.

This leaves just `round` for now, which does not have a WASM
instruction. For now, this operation is not supported. We could
potentially include the version from musl libc (used for wasi-libc).